ScrollLock: Covert component to TypeScript
